Sort some TODOs out
