+++ TODO
@@ -1,0 +1,6 @@
+TODO: add sending metrics
+TODO: check all defers to be sure that error is handled
+TODO: check all fmt.Scan to be sure that error is handled
+TODO: replace all user input with command line args or better with environment variables
+TODO: add PosgreSQL to Docker
+TODO: add my service to Docker
