+++ internal/utils/log_grpc_call.go
@@ -5,6 +5,8 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
+// TODO: add new func which takes funcName, req, res, err, kafkaErr and probably promErr
+
 // LogGrpcCall logs gRPC calls
 func LogGrpcCall(funcName string, req, res interface{}, err *error) {
